harfbuzz/langs: fix max implementation
This commit update the max() helper function to both
compile on 32-bit machines and use the lowest legal
int value on the current architecture as the initial
max value. If the previous min value was chosen for
a specific reason (I was unable to intuit why it was chosen)
this fix may be incorrect.

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>